refactor: seat service 구조 리팩토링
