+++ src/main/java/com/example/book_your_seat/seat/redis/SeatRedisService.java
@@ -1,4 +1,4 @@
-package com.example.book_your_seat.seat.service.redis;
+package com.example.book_your_seat.seat.redis;
 
 import com.example.book_your_seat.reservation.contorller.dto.PaymentRequest;
 import com.example.book_your_seat.seat.domain.Seat;
