+++ src/main/java/com/example/book_your_seat/seat/redis/impl/SeatRedisServiceImpl.java
@@ -1,9 +1,9 @@
-package com.example.book_your_seat.seat.service.redis.impl;
+package com.example.book_your_seat.seat.redis.impl;
 
 
 import com.example.book_your_seat.reservation.contorller.dto.PaymentRequest;
 import com.example.book_your_seat.seat.domain.Seat;
-import com.example.book_your_seat.seat.service.redis.SeatRedisService;
+import com.example.book_your_seat.seat.redis.SeatRedisService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
