+++ src/test/java/com/example/book_your_seat/service/seat/RedisMessageTest.java
@@ -3,8 +3,8 @@ package com.example.book_your_seat.service.seat;
 import com.example.book_your_seat.IntegralTestSupport;
 import com.example.book_your_seat.seat.domain.Seat;
 import com.example.book_your_seat.seat.repository.SeatRepository;
-import com.example.book_your_seat.seat.service.redis.RedisExpiredListener;
-import com.example.book_your_seat.seat.service.redis.SeatRedisService;
+import com.example.book_your_seat.seat.redis.RedisExpiredListener;
+import com.example.book_your_seat.seat.redis.SeatRedisService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
